Move renamePackageOnDisk test to separate file
